MDEV-8302: Duplicate key with parallel replication
Intermediate patch to try to track down the root cause of the problem by
running a custom binary on the user's actual load.

This patch adds some printouts in the error log, all prefixed with
"MDEV8302:", that will hopefully trigger when the error condition occurs and
give more information about what is happening.

In addition, this patch implements what might or might not be a fix for the
error (along with a printout to show if the fix actually triggers).

The fix is to add a check for thd->killed just before a transaction does
mark_start_commit(). This could help reduce the chance of a transaction
detecting a deadlock kill only after running mark_start_commit(), which
might be a source of the original duplicate key problem (though no concrete
sequence of events is currently known to be able to cause this condition).

Finally, this patch marks the generated binaries with a -mdev8302a version
postfix to help identify them.