Fix for Spider bug MDEV-10991 that results in a server crash.
The crash occurs due to code that is #ifdef'd out that pertains to the
use of Spider with an Oracle back-end.  Enabling this code eliminates
the crash.  However, that code doesn't work well and has clearly not
been tested.  We don't want to support an Oracle back-end for Spider,
especially using that code.  I have therefore removed that code.  I have
also removed all of the Spider Oracle tests.