GCF-1058 MTR test galera.MW-86 fails on repeated runs
The reason for the galera.MW-86 failure is not fully understood. A
slightly modified test that fails more easily is planted in
galera.MW-86-fail. It fails when run as:

./mtr --repeat=1000 galera.MW-86-fail

with a result set difference:

 SELECT @@debug_sync;
 @@debug_sync
-ON - signals: ''
+ON - signals: 'signal.wsrep_apply_cb'

meaning that it leaves a pending signal to be unexpectedly consumed by a
subsequent iteration of the test (or another test that uses the same
signal name).

Split galera.MW-86 in two tests as that silences the problem - no
leftover signals are left:

./mtr --no-reorder $(for i in {1..1000} ; do echo galera.MW-86-wait1 galera.MW-86-wait8 ; done)