Mariabackup: Write a dummy empty redo log after --prepare
If --prepare does not create a redo log file, there is a potential
problem: Users could copy the prepared files to a data directory
without deleting pre-existing redo log files. It is safer to create
a dummy redo log file. We cannot use the redo log resizing mechanism
for that, because restoring incremental backups would be broken if
the LSN was advanced if --prepare is invoked on this dummy log file.

InnoDB startup will detect this special dummy file at server startup
and treat it as if no redo log file was present.