MW-388 Move call to after_command() hook to mysql_execute_command()
This implements a 4.x specific fix for MW-388. In this branch the
fix is simply moving call after_command hook to mysql_execute_command().
The reason to move the call is because after_command hook was not
executed within stored procedures.Stored procedures go through a
different parsing path, which missed the call to after_command hook.