GCF-1084 MTR test galera.GCF-1081 fails sporadically
This patch adds a sync point suitable for blocking the wsrep applier
after it BF aborted a victim from within InnoDB. The new sync point
is used to fix test GCF-1081, which attempted to cause a replay of a
Stored Procedure. However, the test was not fully deterministic in
that the applier did not wait long enough until BF abort actually
happened. Because of that the test would fail sporadically, by allowing
executions where the SP committed, without the need of replaying it.