GCF-1096 Fix assertion `!flags.finalized'
The assertion triggers on innodb tests innodb-alter and innodb_bug54044,
with wsrep loaded and binlog_format=STATEMENT.
The problem happens during SELECT phase of CTAS, when when wsrep_pre_commit()
calls wsrep_prepare_data_for_replication(), which fails with
ER_ERROR_DURING_COMMIT due to empty rbr buffer.
The fix here is to strengthen the condition in wsrep_run_hook(), so that
hooks are not triggered during SELECT phase.