The slave io thread does not conduct integrity check for row-based events. Specifically it tolerates missed STMT_END flag marked terminal event that conclude a sequence of "elementary" "block" Rows-events. Failure to react on a potentially missed event can confuse the applier thread in various ways.
The slave io is made in this patch to track the STMT_END status.
Whenever at a first event following a sequence of Rows events the IO thread
finds out that a preceding Rows event did not actually had the flag, an
explicit error is issued.

Replication can be resumed as a provided test demonstrates.

Note that currently the row-based group integrity check is limited
to the version 1 rows event which are generated by Mariadb masters.