MDEV-15016: multiple page cleaner threads use a lot of CPU on idle server
Problem was that when number of page cleaner threads was increased
we sent a event is_requested indicating there is "work" to be done
when there was no work to be done and workers were left on busy loop.
When number of page cleaner threads is decreased we need to sent
is_requested event to indicate that there is "work" to be done.

buf_flush_set_page_cleaner_thread_cnt
        When number of page cleaner threads decrease we send
        is_requested event i.e. this is similar as in shutdown
        there is "work" to be done. We wait these threads
        to exit on is_started event. Page cleaner threads that
        id is less than n_workers will send is_started event
        before decreasing number of running threads and exit.

Manually tested on both (1) idle system and (2) on system with sysbench
workload by increasing and decreasing number of page cleaner threads
and monitoring mysqld CPU usage