MDEV-16307: Incorrect results when using BNLH join instead of BNL join with views
In this issue we are using derived_with_keys optimization and we are using these keys to do a hash join which is incorrect.
We cannot create keys for BLOBS or TEXT columns as only specified length indexes are allowed, so we cannot allow creation of keys
on derived tables if the keyparts have BLOBS or TEXT columns.